requirements.txt
flask
gunicorn
werkzeug
python-pptx
pillow
